refact(redux): reorganize files
