+++ src/pages/DiscordLoginCallback.tsx
@@ -2,7 +2,7 @@ import { Fragment, useEffect, useMemo } from "react";
 import { useNavigate } from "react-router-dom";
 
 import PageTitle from "../components/common/PageTitle";
-import { useLogin } from "../repository/api/useLogin";
+import { useLogin } from "../repository/redux/dispatchs/useLogin";
 import { ExternalRoutesURLs } from "../router/router";
 import useQuery from "../router/useQuery";
