+++ src/repository/commands/useGetOrFetchGuilds.ts
@@ -1,7 +1,6 @@
-import { useFetchGuildsQuery } from "../redux/api";
+import { useFetchGuildsQuery } from "../redux/api/api";
 
 export function useGetOrFetchGuilds() {
   const { data, error, isLoading } = useFetchGuildsQuery();
-
   return { guilds: data, error, loading: isLoading };
 }
