+++ src/repository/commands/useGetOrFetchMe.ts
@@ -1,7 +1,6 @@
-import { useFetchMeQuery } from "../redux/api";
+import { useFetchMeQuery } from "../redux/api/api";
 
 export function useGetorFetchMe() {
   const { data, error, isLoading } = useFetchMeQuery();
-
   return { user: data, error, loading: isLoading };
 }
