+++ src/repository/types/guild.d.ts
@@ -1,0 +1,11 @@
+export interface GuildData {
+  id: string;
+  name: string;
+  icon: string | null;
+  banner: string | null;
+  splash: string | null;
+  owner: boolean;
+  isAdmin: boolean;
+  permissions: string;
+  features: string[];
+}
