+++ src/repository/types/users.d.ts
@@ -4,14 +4,3 @@ export interface AuthenticatedUserObject {
   globalName: string | null;
   avatar: string;
 }
-
-export interface RankedPlayer {
-    ranking: number;
-    user_id: string;
-    xp: number;
-    level: number;
-    xp_to_current_level: number;
-    xp_to_next_level: number;
-    username: string | null;
-    avatar: string;
-}
